Finished 1st round of research.
1. Picked 4 models (K-Means isn't a model) so might try KNN.
2. Tried tuning those models a bit.
3. Used Feature Selection to reduce the amount of features.
4. Used PCA to lessen that even more
5. Removed Ties from Data

Need to start back through and tighten all of these up.  Also need to start with the original training data outputted and figure something out about home/away stats.  If it's better to just use home or away depending on where the current team is playing.  Also to test which features I think are important compared with Feature Selection and maybe even try some random combinations.

Need to also correct the accuracy output so we know the above trials are correct.